+++ build_runner/lib/src/build_plan/package_graph.dart
@@ -155,7 +155,34 @@ class PackageGraph implements AssetPathProvider {
         ),
       );
     }
-    return PackageGraph._(rootNode, nodes);
+
+    // A workspace can have packages that are not depended on by [rootPackage].
+    // Compute transitive dependencies and filter to that.
+    Set<PackageNode>? usedNodes;
+    if (workspacePath != null) {
+      usedNodes = {rootNode};
+      final discoveredNodes = <PackageNode>{rootNode};
+      while (discoveredNodes.isNotEmpty) {
+        final discoveredNodesList = discoveredNodes.toList();
+        discoveredNodes.clear();
+        for (final node in discoveredNodesList) {
+          for (final dep in node.dependencies) {
+            if (usedNodes.add(dep)) {
+              discoveredNodes.add(dep);
+            }
+          }
+        }
+      }
+    }
+
+    return PackageGraph._(
+      rootNode,
+      usedNodes == null
+          ? nodes
+          : Map.fromEntries(
+            nodes.entries.where((e) => usedNodes!.contains(e.value)),
+          ),
+    );
   }
 
   /// Creates a [PackageGraph] for the package in which you are currently